variable size and packing optimization

+++ contracts/test/MockTimeMoodyStakingReward.sol
@@ -3,21 +3,21 @@ pragma solidity ^0.5.16;
 import "../MoodyStakingReward.sol";
 
 contract MockTimeMoodyStakingReward is MoodyStakingReward {
-    uint public time;
+    uint32 public time;
 
     constructor(
         IERC20 rewardToken_,
         IERC20 stakedToken_,
-        uint startingSecond_,
-        uint endingSecond_,
-        uint rewardAmountPerSecond_
+        uint64 startingSecond_,
+        uint64 endingSecond_,
+        uint96 rewardAmountPerSecond_
     ) MoodyStakingReward(rewardToken_, stakedToken_, startingSecond_, endingSecond_, rewardAmountPerSecond_) public {}
 
-    function currentTimestamp() view internal returns (uint) {
+    function currentTimestamp() view internal returns (uint32) {
         return time;
     }
 
-    function setTime(uint time_) public {
+    function setTime(uint32 time_) public {
         time = time_;
     }
 }
